creates __deprecated__ dir and moves textbox
