+++ src/__deprecated__/components/textbox/documentation/info.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { StoryHeader, StoryCode, StoryCodeBlock } from '../../../../.storybook/style/storybook-info.styles';
+import { StoryHeader, StoryCode, StoryCodeBlock } from '../../../../../.storybook/style/storybook-info.styles';
 
 const info = (
   <div>
